simulator updates
git-svn-id: svn://cs2955/var/svn/client/trunk@357 729a0f52-b9d0-41be-9dfd-8227063685e9

+++ src/newmodels/usermodel/AbstractUserModel.java
@@ -18,6 +18,6 @@ public abstract class AbstractUserModel extends AbstractModel {
 	
 	public abstract boolean updateModel(QueryReport queryReport, SalesReport salesReport);
 	
-	public abstract double getPrediction(Product product, UserState userState);
+	public abstract int getPrediction(Product product, UserState userState);
 
 }
